requirements.txt
requests==2.22.0
python-dotenv==0.10.2
instabot==0.42.0
